+++ src/utils/errorOutput.ts
@@ -2,6 +2,7 @@ type DataType = {
   detail: string;
   details: string;
   field_errors: string[];
+  non_field_errors: string[];
 };
 type ResponseType = {
   data: DataType;
@@ -22,6 +23,10 @@ const errorOutput = (error: ErrorType | any): string | ErrorType | any => {
   if (error?.response?.data.field_errors) {
     const err = error.response.data.field_errors;
     return Object.values(err).join(' ');
+  }
+  if (error?.response?.data.non_field_errors) {
+    const err = error.response.data.non_field_errors;
+    return err;
   }
   return error;
 };